Move PluginSetting.cs to new folder/namespace (more to come for this ns); No functional code changes.
