+++ MSFSTouchPortalPlugin/Interfaces/IReflectionService.cs
@@ -1,4 +1,5 @@
 ﻿using MSFSTouchPortalPlugin.Constants;
+using MSFSTouchPortalPlugin.Types;
 using System;
 using System.Collections.Generic;
 
@@ -7,6 +8,6 @@ namespace MSFSTouchPortalPlugin.Interfaces {
     Dictionary<string, Enum> GetInternalEvents();
     Dictionary<string, Enum> GetActionEvents();
     Dictionary<Definition, SimVarItem> GetStates();
-    Dictionary<string, Objects.Plugin.PluginSetting> GetSettings();
+    Dictionary<string, PluginSetting> GetSettings();
   }
 }
